+++ CHANGELOG.md
@@ -83,6 +83,7 @@
 - Enh #84: Add `HasDy` trait and corresponding tests for managing SVG `dy` attribute (@terabytesoftw)
 - Enh #85: Add `HasFontFamily` trait and corresponding tests for managing SVG `font-family` attribute (@terabytesoftw)
 - Bug #86: Standardize all documentation in related classes (@terabytesoftw)
+- Enh #87: Add `HasFontSize` trait and corresponding tests for managing SVG `font-size` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024